refactor: Extract read_parquet_file test helper function to iox_tests::utils
